+++ src/main/java/com/enjin/enjincoin/sdk/model/service/requests/data/CreateTradeData.java
@@ -3,7 +3,6 @@ package com.enjin.enjincoin.sdk.model.service.requests.data;
 import com.google.gson.annotations.SerializedName;
 import lombok.Builder;
 
-import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -57,6 +56,6 @@ public class CreateTradeData {
      * @return the builder.
      */
     @SerializedName("second_party_identity_id")
-    private BigInteger secondPartyIdentityId;
+    private Integer secondPartyIdentityId;
 
 }
