+++ src/main/java/com/enjin/enjincoin/sdk/model/service/requests/data/SendTokenData.java
@@ -3,8 +3,6 @@ package com.enjin.enjincoin.sdk.model.service.requests.data;
 import com.google.gson.annotations.SerializedName;
 import lombok.Builder;
 
-import java.math.BigInteger;
-
 /**
  * A data model for the Send Token request type.
  *
@@ -56,7 +54,7 @@ public class SendTokenData {
      * @return the builder.
      */
     @SerializedName("recipient_identity_id")
-    private BigInteger recipientIdentityId;
+    private Integer recipientIdentityId;
 
     /**
      * The amount to send.
